Update Command Input, Name Tags Before Rebuilding
TL;DR: Make Command Idiot Proof

+++ TagTool/Commands/MtnDewIt/BaseCache/BaseCacheCommand.Cache.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using TagTool.Cache;
 using TagTool.Cache.Gen3;
 using TagTool.Cache.HaloOnline;
@@ -194,6 +195,8 @@ namespace TagTool.Commands.Tags
             CopiedResources.Clear();
 
             var destDirectory = new DirectoryInfo(destCacheDirectory);
+
+            EmptyDirectory(destDirectory);
 
             CreateTagCache(destDirectory);
 
@@ -545,7 +548,35 @@ namespace TagTool.Commands.Tags
         {
             Directory.CreateDirectory($@"{path}\fonts");
 
-            File.Copy($@"{Program.TagToolDirectory}\Tools\BaseCache\Fonts\Upscaled\font_package.bin", $@"{path}\fonts\font_package.bin");
+            if (!File.Exists($@"{path}\fonts\font_package.bin"))
+            {
+                File.Copy($@"{Program.TagToolDirectory}\Tools\BaseCache\Fonts\Upscaled\font_package.bin", $@"{path}\fonts\font_package.bin");
+            }
+            else
+            {
+                new TagToolWarning($@"Font Package Detected in Specified Directory! Replacing Anyway.");
+                File.Copy($@"{Program.TagToolDirectory}\Tools\BaseCache\Fonts\Upscaled\font_package.bin", $@"{path}\fonts\font_package.bin", true);
+            }
+        }
+
+        public void EmptyDirectory(DirectoryInfo directoryInfo)
+        {
+            string[] remove = { "map", "dat", "csv" };
+
+            if (directoryInfo.GetFiles().Any(x => x.FullName.EndsWith(".map") || x.FullName.EndsWith(".dat") || x.FullName.EndsWith(".csv")))
+            {
+                new TagToolWarning($@"Cache Detected in Specified Directory! Replacing Anyway.");
+
+                foreach (string fileType in remove)
+                {
+                    FileInfo[] files = directoryInfo.GetFiles("*." + fileType);
+
+                    foreach (FileInfo file in files)
+                    {
+                        file.Delete();
+                    }
+                }
+            }
         }
 
         public void SetCacheVersion(GameCacheHaloOnline cache, CacheVersion version)
